+++ cmds/library.py
@@ -25,8 +25,8 @@ import os,pickle,time,traceback
 
 import smbc
 
-from decorators import *
-import util
+from lib.decorators import *
+import lib.util as util
 
 @botconf
 def conf(bot):